change image url for higher res images. refactor location of components
