+++ app/serializers/team_serializer.rb
@@ -1,3 +1,3 @@
 class TeamSerializer < ActiveModel::Serializer
-  attributes *Team.column_names, :record
+  attributes *Team.column_names, :record, :sport
 end
